component 분리
