+++ features/retrospect/src/main/java/com/depromeet/team5/features/retrospect/screen/RetrospectScreen.kt
@@ -5,7 +5,6 @@ import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -18,7 +17,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.ButtonDefaults
@@ -26,32 +24,26 @@ import androidx.compose.material3.Switch
 import androidx.compose.material3.SwitchDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.depromeet.team5.features.retrospect.R
-import com.depromeet.team5.features.retrospect.screen.annotation.CurrencyUnit
 import com.depromeet.team5.features.retrospect.screen.annotation.KOREAN
 import com.depromeet.team5.features.retrospect.screen.annotation.PERCENT
+import com.depromeet.team5.features.retrospect.screen.component.HedgeTextField
+import com.depromeet.team5.features.retrospect.screen.component.HedgeTopbar
 import kotlinx.coroutines.delay
 
 
@@ -216,35 +208,6 @@ private fun RetrospectScreen(
 }
 
 @Composable
-private fun HedgeTopbar(
-    modifier: Modifier = Modifier,
-    onBackPressed: () -> Unit
-) {
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(44.dp)
-            .background(color = Color(0xFFF3F4F6)),
-        contentAlignment = Alignment.CenterStart
-    ) {
-
-        Box(
-            modifier = Modifier
-                .size(40.dp)
-                .padding(start = 12.dp),
-            contentAlignment = Alignment.Center
-        ) {
-            Image(
-                modifier = Modifier
-                    .size(24.dp),
-                painter = painterResource(R.drawable.icon_arrow_left),
-                contentDescription = null
-            )
-        }
-    }
-}
-
-@Composable
 private fun CompanyTitle(
     modifier: Modifier = Modifier
 ) {
@@ -262,86 +225,8 @@ private fun CompanyTitle(
             modifier = Modifier.padding(start = 7.dp),
             text = "삼성전자"
         )
-
     }
 }
-
-@Composable
-private fun HedgeTextField(
-    modifier: Modifier = Modifier,
-    label: String,
-    placeholder: String,
-    unit: CurrencyUnit? = null
-) {
-    var borderColor by remember { mutableIntStateOf(R.color.brand500) }
-    var inputText by remember { mutableStateOf("") }
-
-    Box(
-        modifier = modifier
-            .padding(1.dp)
-            .fillMaxWidth()
-            .wrapContentHeight()
-            .background(
-                color = colorResource(R.color.white),
-                shape = RoundedCornerShape(16.dp)
-            )
-            .border(1.5.dp, colorResource(borderColor), RoundedCornerShape(16.dp))
-            .onFocusChanged { focusState ->
-                borderColor = if (focusState.hasFocus) R.color.gray900
-                else R.color.white
-
-            }
-            .padding(vertical = 14.dp, horizontal = 20.dp)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            TextField(
-                modifier = Modifier
-                    .wrapContentSize()
-                    .padding(top = 4.dp),
-                value = inputText,
-                onValueChange = { inputText = it },
-                visualTransformation = if (unit != null) {
-                    CurrencyVisualTransformation(unit)
-                } else {
-                    VisualTransformation.None
-                },
-                label = {
-                    Text(
-                        text = label,
-                        fontWeight = FontWeight.SemiBold,
-                        fontSize = 13.sp
-                    )
-                },
-                placeholder = { Text(placeholder) },
-                colors = TextFieldDefaults.colors(
-                    focusedContainerColor = Color.Transparent,
-                    unfocusedContainerColor = Color.Transparent,
-                    disabledContainerColor = Color.Transparent,
-
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent,
-                    disabledIndicatorColor = Color.Transparent,
-                    errorIndicatorColor = Color.Transparent,
-
-                    focusedLabelColor = colorResource(R.color.gray900),
-                    unfocusedLabelColor = colorResource(R.color.gray500),
-
-                    focusedPlaceholderColor = colorResource(R.color.gray400),
-
-                    focusedTextColor = colorResource(R.color.gray900),
-
-                    cursorColor = Color.Transparent,
-                    errorCursorColor = Color.Transparent,
-                )
-            )
-        }
-    }
-}
-
 
 @Preview
 @Composable
@@ -359,20 +244,9 @@ private fun CompanyTitlePreview() {
 
 @Preview
 @Composable
-private fun HedgeTextFieldPreview() {
-    HedgeTextField(
-        label = "매도가",
-        placeholder = "매도 가격"
-    )
-
-}
-
-@Preview
-@Composable
 private fun RetrospectRoutePreview() {
     RetrospectRoute(
         modifier = Modifier,
         onBackPressed = {}
     )
 }
-
